Crew_Turtles/CrewAI_Turtles_Crew/pyproject.toml
[tool.poetry]
name = "crewai-turtle-crew"
version = "0.1.0"
description = "CrewAI Turtle Crew"
authors = ["Rizwan A. & Ahmed A."]

[tool.poetry.dependencies]
python = ">=3.10.0,<3.12"
python-dotenv = "1.0.0"
unstructured = "==0.10.25"
langchain-openai = "^0.0.5"
decouple = "^0.0.7"
crewai = "^0.22.5"
crewai-tools = "^0.1.1"
langchain-community = "^0.0.29"
pydantic = "^2.6.4"
streamlit = "^1.32.2"

[tool.pyright]
# https://github.com/microsoft/pyright/blob/main/docs/configuration.md
useLibraryCodeForTypes = true
exclude = [".cache"]

[tool.ruff]
# https://beta.ruff.rs/docs/configuration/
select = ['E', 'W', 'F', 'I', 'B', 'C4', 'ARG', 'SIM']
ignore = ['W291', 'W292', 'W293']

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"